Add possibility to generate difference tokens for RUD operations

+++ app/controllers/Page/PageDelete.js
@@ -2,8 +2,13 @@ const {PageBase} = require('./PageBase');
 const he = require('he');
 const db = require('../../services/DBService');
 const log = require('../../singleton/logger');
+const {OPERATIONS} = require('../../constants');
 
 class PageDelete extends PageBase{
+    /**
+     * @param eid
+     * @param access_token
+     */
     constructor({eid, access_token}) {
         super();
 
@@ -11,11 +16,15 @@ class PageDelete extends PageBase{
         this.access_token = access_token || null;
     }
 
+    /**
+     * Remove saved data from DB
+     * @returns {Promise<void>}
+     */
     async remove() {
         log.debug('Delete page with eid: %s', this.eid);
 
-        await this._DBQueryHandler(async () => {
-            await this._checkAbleToModify();
+        await this.DBQueryHandler(async () => {
+            await this.checkAccessibility(OPERATIONS.D);
             await db.removePage(this.eid);
         });
